[REFACT] 유저 업적 관련 서비스 분리
